+++ src/components/article-params-form/ArticleParamsForm.module.scss
@@ -31,13 +31,3 @@
 	justify-content: space-between;
 	margin-top: auto;
 }
-
-.title {
-	color: #000000;
-	font-weight: 800;
-	font-size: 31px;
-	font-family: 'Open Sans', sans-serif;
-	line-height: normal;
-	letter-spacing: 0%;
-	text-transform: uppercase;
-}
