+++ src/components/article/Article.tsx
@@ -1,5 +1,6 @@
 import clsx from 'clsx';
 
+// Сообщаем вебпаку, что этот файл использует это изображение.
 import plane from 'src/images/plane.png';
 import { Text } from 'components/text';
